Made stimulus larger
also changed it so that when the text version of the answer (var: keyboard_response) doesn't crash the app when no audio was understood. And the keyboard_response is checked through 'answer in keyboard_response' instead of 'keyboard_response == answer" to ensure that saying the answer twice doesn't get marked as incorrect.